web-platform_22_10_dev/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-grappelli==3.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
frozenlist==1.3.1
idna==3.4
lxml==4.9.1
multidict==6.0.2
openpyxl==3.0.10
Pillow==9.2.0
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.5
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
yarl==1.8.1
